# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kiroulou API Documentation",
        "description": "Documentation de l'API utilisée par l'application mobile KiRoulOu, faite avec Laravel",
        "contact": {
            "email": "bezedache29@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "https://localhost/api",
            "description": "Server API KiRoulOu"
        }
    ],
    "paths": {
...
        "/clubs/{club_id}/posts?page={page}": {
            "get": {
                "tags": [
                    "Clubs"
                ],
                "summary": "All club posts",
                "operationId": "90bf0f24c2280417d6f6e01aca432e3e",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/club_id"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ClubPostCounts"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/clubs/{club_id}/posts": {
            "post": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Create club post",
                "description": "Ajout d'un article par un club",
                "operationId": "d4fd58758e03d42acf3c6461f094dab8",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/club_id"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/AddClubPost"
                },
                "responses": {
                    "201": {
                        "description": "Article d'un club créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "post created"
                                        },
                                        "post": {
                                            "$ref": "#/components/schemas/ClubPostSimple"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/clubs/{club_id}/posts/{post_id}/comments": {
            "post": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Create club post comment",
                "description": "Ajout d'un commentaire à un article de club",
                "operationId": "89560aff7aa2df8192802f4bcbee53bf",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/club_id"
                    },
                    {
                        "$ref": "#/components/parameters/post_id"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/AddClubPostComment"
                },
                "responses": {
                    "201": {
                        "description": "Commentaire d'article de club créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "comment created"
                                        },
                                        "comments": {
                                            "$ref": "#/components/schemas/ClubPostComment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/clubs/{club_id}/posts/{post_id}/comments?page={page}": {
            "get": {
                "tags": [
                    "Clubs"
                ],
                "summary": "All club post comments",
                "operationId": "6229ba3dc58ead857a017856d61694cd",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/club_id"
                    },
                    {
                        "$ref": "#/components/parameters/post_id"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ClubPostComment"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/clubs/{club_id}/posts/{post_id}/likeOrUnlike": {
            "post": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Like or unlike club post",
                "description": "Aimer ou ne plus aimer un article d'un club",
                "operationId": "6f65932e1b05133f9d29c688eb24f368",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/club_id"
                    },
                    {
                        "$ref": "#/components/parameters/post_id"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Article d'un user créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "liked"
                                        },
                                        "post": {
                                            "$ref": "#/components/schemas/ClubPostLikes"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClubPostSimple": {
                "description": "Article d'un club",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "b7d15d52-0458-4a39-a59e-a7a370e7f31c"
                    },
                    "title": {
                        "type": "string",
                        "example": "Mon super titre"
                    },
                    "description": {
                        "type": "string",
                        "example": "Ma super description d'article"
                    },
                    "club_id": {
                        "type": "number",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2022-08-15 14:25:01"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": null
                    }
                },
                "type": "object"
            },
...
            "ClubPostCounts": {
                "description": "Article d'un club avec count likes et commentaires",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ClubPostSimple"
                    },
                    {
                        "properties": {
                            "post_likes_count": {
                                "type": "number",
                                "example": 36
                            },
                            "post_comments_count": {
                                "type": "number",
                                "example": 27
                            }
                        }
                    }
                ]
            },
...
            "ClubPostLikes": {
                "description": "Article d'un club avec count likes",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ClubPostSimple"
                    },
                    {
                        "properties": {
                            "post_likes_count": {
                                "type": "number",
                                "example": 36
                            }
                        }
                    }
                ]
            },
            "ClubPostComment": {
                "description": "Commentaire d'un article d'un club par un user",
                "properties": {
                    "id": {
                        "type": "number",
                        "example": 1
                    },
                    "user_id": {
                        "type": "number",
                        "example": 10005
                    },
                    "club_post_id": {
                        "type": "string",
                        "example": "37e918d3-ece7-4f45-a54d-838d951c2c8d"
                    },
                    "message": {
                        "type": "string",
                        "example": "Mon super commentaire d'un article du club"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2022-08-15 14:25:01"
                    },
                    "user_name": {
                        "description": "Nom et Prénom du user ou son email si non renseigné",
                        "type": "string",
                        "example": "Simon Strueux"
                    },
                    "user_club_name": {
                        "description": "Nom du club du user ou la ville du user si non renseigné",
                        "type": "string",
                        "example": "Côte Des Légendes VTT"
                    },
                    "user_avatar_name": {
                        "description": "Avatar du user ou celui par default si non renseigné",
                        "type": "string",
                        "example": "mon-avatar.png"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "UnprocessableEntity": {
                "description": "Ressource non traitée",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "array",
                                    "items": {},
                                    "example": [
                                        "L'adresse email est obligatoire"
                                    ]
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "NotFound": {
                "description": "Ressource introuvable",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Resource not found"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
...
            "club_id": {
                "name": "club_id",
                "in": "path",
                "description": "ID du club",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
            "post_id": {
                "name": "post_id",
                "in": "path",
                "description": "ID du post (uuid)",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
...
            "page": {
                "name": "page",
                "in": "path",
                "description": "Numéro de la page",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
        },
        "requestBodies": {
...
            "AddClubPost": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "title, description"
                            ],
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "example": "Mon super titre"
                                },
                                "description": {
                                    "type": "string",
                                    "example": "Ma super description"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "AddClubPostComment": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "message"
                            ],
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Mon super commentaire"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "S'authentifier avec email & password pour récupérer un token d'authentification",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}